refactor: merge handlers having views
