--- app/src/main/java/com/techcourse/controller/RegisterViewController.java
@@ -1,16 +1,0 @@
-package com.techcourse.controller;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.web.annotation.Controller;
-import nextstep.web.annotation.RequestMapping;
-import nextstep.web.support.RequestMethod;
-
-@Controller
-public class RegisterViewController {
-
-    @RequestMapping(value = "/register/view", method = RequestMethod.GET)
-    public String handleRegisterView(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        return "/register.jsp";
-    }
-}
